+++ src/app/components/listing/inspect/inspect.controller.js
@@ -32,9 +32,8 @@
 
         vm.certificationStatus = utilService.certificationStatus;
         vm.isBlank = utilService.isBlank;
-        vm.getAttestationStringForCurrentSystemDeveloper = getAttestationStringForCurrentSystemDeveloper;
+        vm.getAttestationForCurrentSystemDeveloper = getAttestationForCurrentSystemDeveloper;
         vm.populateDeveloperSystemRequirements = populateDeveloperSystemRequirements;
-        vm.isTransparencyAttestationViewable = isTransparencyAttestationViewable;
 
         vm.isOn = featureFlags.isOn;
 
@@ -97,13 +96,20 @@
                     name: vm.cp.developer.name,
                     status: vm.developer.status,
                     statusEvents: vm.developer.statusEvents,
-                    transparencyAttestations: [{acbId: vm.cp.certifyingBody.id, acbName: vm.cp.certifyingBody.name, attestation: vm.cp.transparencyAttestation}],
+                    transparencyAttestations: [{
+                        acbId: vm.cp.certifyingBody.id,
+                        acbName: vm.cp.certifyingBody.name,
+                        attestation: { transparencyAttestation: vm.cp.transparencyAttestation.transparencyAttestation },
+                    }],
                     website: vm.cp.developer.website,
                 },
                 developerIds: [vm.cp.developer.developerId],
             };
             if (!dev.developer.address.country) {
                 dev.developer.address.country = 'USA';
+            }
+            if (vm.isOn('effective-rule-date-plus-one-week')) {
+                delete dev.developer.transparencyAttestations;
             }
             networkService.updateDeveloper(dev)
                 .then(function () {
@@ -239,7 +245,8 @@
                 && !vm.isBlank(vm.developer.contact.phoneNumber))
                 && (vm.developer.address && !vm.isBlank(vm.developer.address.line1) && !vm.isBlank(vm.developer.address.city)
                 && !vm.isBlank(vm.developer.address.state) && !vm.isBlank(vm.developer.address.zipcode))
-                && (!vm.isBlank(vm.getAttestationStringForCurrentSystemDeveloper())))) {
+                && (vm.getAttestationForCurrentSystemDeveloper())
+                && (!vm.isBlank(vm.getAttestationForCurrentSystemDeveloper().transparencyAttestation)))) {
                 return true;
             }
             vm.populateDeveloperSystemRequirements();
@@ -277,13 +284,13 @@
                     vm.systemRequirements.push('None of the required developer address information'
                         + EXISTS_MSG + PLEASE_SAVE_MSG);
                 }
-                if (vm.isBlank(vm.getAttestationStringForCurrentSystemDeveloper())) {
+                if (!vm.getAttestationForCurrentSystemDeveloper() || vm.isBlank(vm.getAttestationForCurrentSystemDeveloper().transparencyAttestation)) {
                     vm.systemRequirements.push('A transparency attestation' + DOES_NOT_EXIST_MSG + PLEASE_SAVE_MSG);
                 }
             }
         }
 
-        function getAttestationStringForCurrentSystemDeveloper () {
+        function getAttestationForCurrentSystemDeveloper () {
             if (vm.developer && vm.developer.transparencyAttestations) {
                 let matchingAttestationObj = vm.developer.transparencyAttestations.find(function (curAttestationObj) {
                     return curAttestationObj.acbName === vm.cp.certifyingBody.name;
@@ -291,13 +298,6 @@
                 return matchingAttestationObj ? matchingAttestationObj.attestation : undefined;
             }
             return null;
-        }
-
-        function isTransparencyAttestationViewable () {
-            if (vm.isOn('effective-rule-date-plus-one-week')) {
-                return !vm.isAcbAdmin;
-            }
-            return true;
         }
 
         function cancel () {